+++ hydromt/models/model_api.py
@@ -3,7 +3,7 @@
 
 from abc import ABCMeta
 import os, glob
-from os.path import join, isdir, isfile, abspath, dirname, basename
+from os.path import join, isdir, isfile, abspath, dirname, basename, isabs
 import xarray as xr
 import numpy as np
 import geopandas as gpd
@@ -26,80 +26,6 @@ __all__ = ["Model"]
 logger = logging.getLogger(__name__)
 
 
-class AuxmapsMixin(object):
-    # mixin class to add an auxiliary maps object
-    # contains maps needed for model building but not model data
-    _API = {
-        "auxmaps": Dict[str, Union[xr.DataArray, xr.Dataset]],
-    }
-
-    def __init__(self, *args, **kwargs) -> None:
-        super().__init__(*args, **kwargs)
-        self._auxmaps = dict()  # dictionary of xr.DataArray and/or xr.Dataset
-
-    # model auxiliary map files
-    @property
-    def auxmaps(self) -> Dict[str, Union[xr.Dataset, xr.DataArray]]:
-        """Auxillary model maps. Returns dict of xarray.DataArray or xarray.Dataset"""
-        if len(self._auxmaps) == 0:
-            if self._read:
-                self.read_auxmaps()
-        return self._auxmaps
-
-    def set_auxmaps(
-        self,
-        data: Union[xr.DataArray, xr.Dataset],
-        name: Optional[str] = None,
-        split_dataset: Optional[bool] = False,
-    ) -> None:
-        """Add auxiliary data to maps.
-
-        Dataset can either be added as is (default) or split into several
-        DataArrays using the split_dataset argument.
-
-        Arguments
-        ---------
-        data: xarray.Dataset or xarray.DataArray
-            New forcing data to add
-        name: str, optional
-            Variable name, only in case data is of type DataArray or if a Dataset is added as is (split_dataset=False).
-        split_dataset: bool, optional
-            If data is a xarray.Dataset, either add it as is to results or split it into several xarray.DataArrays.
-        """
-        data_dict = _check_data(data, name, split_dataset)
-        for name in data_dict:
-            if name in self._auxmaps:
-                self.logger.warning(f"Replacing result: {name}")
-            self._auxmaps[name] = data_dict[name]
-
-    def read_auxmaps(self, fn: str = "auxmaps/*.nc", **kwargs) -> None:
-        """Read auxillary model map at <root>/<fn> and add to maps property
-
-        key-word arguments are passed to :py:func:`xarray.open_dataset`
-
-        Parameters
-        ----------
-        fn : str, optional
-            filename relative to model root, may wildcards, by default "auxmaps/*.nc"
-        """
-        ncs = self._read_nc(fn, **kwargs)
-        for name, ds in ncs.items():
-            self.set_auxmaps(ds, name=name)
-
-    def write_auxmaps(self, fn="auxmaps/{name}.nc", **kwargs) -> None:
-        """Write auxmaps to netcdf file at <root>/<fn>
-
-        key-word arguments are passed to :py:meth:`xarray.Dataset.to_netcdf`
-
-        Parameters
-        ----------
-        fn : str, optional
-            filename relative to model root and should contain a {name} placeholder,
-            by default 'auxmaps/{name}.nc'
-        """
-        self._write_nc(self._auxmaps, fn, **kwargs)
-
-
 class Model(object, metaclass=ABCMeta):
     """General and basic API for models in HydroMT"""
 
@@ -119,6 +45,7 @@ class Model(object, metaclass=ABCMeta):
         "crs": CRS,
         "config": Dict[str, Any],
         "geoms": Dict[str, gpd.GeoDataFrame],
+        "maps": Dict[str, Union[xr.DataArray, xr.Dataset]],
         "forcing": Dict[str, Union[xr.DataArray, xr.Dataset]],
         "region": gpd.GeoDataFrame,
         "results": Dict[str, Union[xr.DataArray, xr.Dataset]],
@@ -130,7 +57,7 @@ class Model(object, metaclass=ABCMeta):
         root: Optional[str] = None,
         mode: Optional[str] = "w",
         config_fn: Optional[str] = None,
-        data_libs: Optional[List[str]] = None,
+        data_libs: Union[List, str] = [],
         logger=logger,
         **artifact_keys,
     ):
@@ -148,12 +75,13 @@ class Model(object, metaclass=ABCMeta):
         data_libs : List[str], optional
             List of data catalog yaml files, by default None
         """
-        from . import ENTRYPOINTS  # load within method to avoid circular imports
+        from . import MODELS  # avoid circular import
 
         self.logger = logger
-        ep = ENTRYPOINTS.get(self._NAME, None)
-        version = ep.distro.version if ep is not None else ""
-        dist = ep.distro.name if ep is not None else "unknown"
+        dist, version = "unknown", "NA"
+        if self._NAME in MODELS:
+            ep = MODELS[self._NAME]
+            dist, version = ep.distro.name, ep.distro.version
 
         # link to data
         self.data_catalog = DataCatalog(
@@ -163,9 +91,9 @@ class Model(object, metaclass=ABCMeta):
         # placeholders
         # metadata maps that can be at different resolutions #TODO> do we want read/write maps?
         self._config = dict()  # nested dictionary
-        self._geoms = (
-            dict()
-        )  # dictionary of gdp.GeoDataFrame NOTE was staticgeoms in <=v0.5
+        self._maps = dict()  # dictionary of xr.DataArray and/or xr.Dataset
+        # NOTE was staticgeoms in <=v0.5
+        self._geoms = dict()  # dictionary of gdp.GeoDataFrame
         self._forcing = dict()  # dictionary of xr.DataArray and/or xr.Dataset
         self._states = dict()  # dictionary of xr.DataArray and/or xr.Dataset
         self._results = dict()  # dictionary of xr.DataArray and/or xr.Dataset
@@ -397,7 +325,7 @@ class Model(object, metaclass=ABCMeta):
             For a complete overview of all region options,
             see :py:function:~hydromt.workflows.basin_mask.parse_region
         hydrography_fn : str
-            Name of data source for basemap parameters.
+            Name of data source for hydrography data.
             FIXME describe data requirements
         basin_index_fn : str
             Name of data source with basin (bounding box) geometries associated with
@@ -465,6 +393,16 @@ class Model(object, metaclass=ABCMeta):
             raise ValueError("Root unknown, use set_root method")
         return self._root
 
+    @property
+    def _assert_write_mode(self):
+        if not self._write:
+            raise IOError("Model opened in read-only mode")
+
+    @property
+    def _assert_read_mode(self):
+        if not self._read:
+            raise IOError("Model opened in write-only mode")
+
     def set_root(self, root: Optional[str], mode: Optional[str] = "w"):
         """Initialize the model root.
 
@@ -498,11 +436,6 @@ class Model(object, metaclass=ABCMeta):
             # check directory
             elif not isdir(self._root):
                 raise IOError(f'model root not found at "{self._root}"')
-            # read hydromt_data yaml file and add to data catalog
-            data_fn = join(self._root, "hydromt_data.yml")
-            if self._read and isfile(data_fn):
-                # read data and mark as used
-                self.data_catalog.from_yml(data_fn, mark_used=True)
             # remove old logging file handler and add new filehandler in root if it does not exist
             has_log_file = False
             log_level = 20  # default, but overwritten by the level of active loggers
@@ -526,6 +459,7 @@ class Model(object, metaclass=ABCMeta):
         components: List = [
             "config",
             "staticmaps",
+            "maps",
             "geoms",
             "forcing",
             "states",
@@ -552,6 +486,7 @@ class Model(object, metaclass=ABCMeta):
         self,
         components: List = [
             "staticmaps",
+            "maps",
             "geoms",
             "forcing",
             "states",
@@ -575,27 +510,50 @@ class Model(object, metaclass=ABCMeta):
             getattr(self, f"write_{component}")()
 
     def write_data_catalog(
-        self, root: Optional[Union[str, Path]] = None, used_only: bool = True
+        self,
+        root: Optional[Union[str, Path]] = None,
+        data_lib_fn: Union[str, Path] = "hydromt_data.yml",
+        used_only: bool = True,
+        append: bool = True,
     ):
-        """Write the data catalog to `hydromt_data.yml`
+        """Write the data catalog to data_lib_fn
 
         Parameters
         ----------
         root: str, Path, optional
             Global root for all relative paths in yaml file.
             If "auto" the data source paths are relative to the yaml output ``path``.
-        used_only: bool
-            If True, export only data entries kept in used_data list.
+        data_lib_fn: str, Path, optional
+            Path of output yml file, absolute or relative to the model root, by default "hydromt_data.yml".
+        used_only: bool, optional
+            If True, export only data entries kept in used_data list. By default True
+        append: bool, optional
+            If True, append to an existing
         """
-        path = join(self.root, "hydromt_data.yml")
-        self.data_catalog.to_yml(path, root=root, used_only=used_only)
+        path = data_lib_fn if isabs(data_lib_fn) else join(self.root, data_lib_fn)
+        cat = DataCatalog(logger=self.logger, fallback_lib=None)
+        # read hydromt_data yaml file and add to data catalog
+        if self._read and isfile(path) and append:
+            cat.from_yml(path)
+        # update data catalog with new used sources
+        source_names = (
+            self.data_catalog._used_data
+            if used_only
+            else list(self.data_catalog.sources.keys())
+        )
+        if len(source_names) > 0:
+            cat.from_dict(self.data_catalog.to_dict(source_names=source_names))
+        if cat.sources:
+            self._assert_write_mode
+            cat.to_yml(path, root=root)
 
     # model configuration
     @property
     def config(self) -> Dict[str, Union[Dict, str]]:
         """Model configuration. Returns a (nested) dictionary"""
+        # initialize default config if in write-mode
         if not self._config:
-            self.read_config()  # initialize default config
+            self.read_config()
         return self._config
 
     def set_config(self, *args):
@@ -715,8 +673,7 @@ class Model(object, metaclass=ABCMeta):
         self, config_name: Optional[str] = None, config_root: Optional[str] = None
     ):
         """Write config to <root/config_fn>"""
-        if not self._write:
-            raise IOError("Model opened in read-only mode")
+        self._assert_write_mode
         if config_name is not None:
             self._config_fn = config_name
         elif self._config_fn is None:
@@ -738,9 +695,8 @@ class Model(object, metaclass=ABCMeta):
             "use the grid property of the GridModel class instead.",
             DeprecationWarning,
         )
-        if len(self._staticmaps) == 0:
-            if self._read:
-                self.read_staticmaps()
+        if len(self._staticmaps) == 0 and self._read:
+            self.read_staticmaps()
         return self._staticmaps
 
     def set_staticmaps(
@@ -805,6 +761,7 @@ class Model(object, metaclass=ABCMeta):
         fn : str, optional
             filename relative to model root, by default "staticmaps/staticmaps.nc"
         """
+        self._assert_read_mode
         for ds in self._read_nc(fn, **kwargs).values():
             self.set_staticmaps(ds)
 
@@ -820,20 +777,221 @@ class Model(object, metaclass=ABCMeta):
         fn : str, optional
             filename relative to model root, by default 'staticmaps/staticmaps.nc'
         """
-        nc_dict = dict()
-        if len(self._staticmaps) > 0:
-            # _write_nc requires dict - use dummy key
-            nc_dict.update({"staticmaps": self._staticmaps})
-        self._write_nc(nc_dict, fn, **kwargs)
+        if len(self._staticmaps) == 0:
+            self.logger.debug("No staticmaps data found, skip writing.")
+        else:
+            self._assert_write_mode
+            # _write_nc requires dict - use dummy 'staticmaps' key
+            nc_dict = {"staticmaps": self._staticmaps}
+            self._write_nc(nc_dict, fn, **kwargs)
+
+    # map files setup methods
+    def setup_maps_from_raster(
+        self,
+        raster_fn: str,
+        variables: Optional[List] = None,
+        fill_method: Optional[str] = None,
+        name: Optional[str] = None,
+        reproject_method: Optional[str] = None,
+        split_dataset: Optional[bool] = True,
+    ) -> List[str]:
+        """
+        This component adds data variable(s) from ``raster_fn`` to maps object.
+
+        If raster is a dataset, all variables will be added unless ``variables`` list is specified.
+
+        Adds model layers:
+
+        * **raster.name** maps: data from raster_fn
+
+        Parameters
+        ----------
+        raster_fn: str
+            Source name of raster data in data_catalog.
+        variables: list, optional
+            List of variables to add to maps from raster_fn. By default all.
+        fill_method : str, optional
+            If specified, fills nodata values using fill_nodata method.
+            Available methods are {'linear', 'nearest', 'cubic', 'rio_idw'}.
+        name: str, optional
+            Name of new maps variable, only in case split_dataset=False.
+        reproject_method: str, optional
+            See rasterio.warp.reproject for existing methods, by default the data is not reprojected (None).
+        split_dataset: bool, optional
+            If data is a xarray.Dataset split it into several xarray.DataArrays (default).
+
+        Returns
+        -------
+        list
+            Names of added model map layers
+        """
+        self.logger.info(f"Preparing maps data from raster source {raster_fn}")
+        # Read raster data and select variables
+        ds = self.data_catalog.get_rasterdataset(
+            raster_fn,
+            geom=self.region,
+            buffer=2,
+            variables=variables,
+            single_var_as_array=False,
+        )
+        # Fill nodata
+        if fill_method is not None:
+            ds = ds.raster.interpolate_na(method=fill_method)
+        # Reprojection
+        if ds.rio.crs != self.crs and reproject_method is not None:
+            ds = ds.raster.reproject(dst_crs=self.crs, method=reproject_method)
+        # Add to maps
+        self.set_maps(ds, name=name, split_dataset=split_dataset)
+
+        return list(ds.data_vars.keys())
+
+    def setup_maps_from_raster_reclass(
+        self,
+        raster_fn: str,
+        reclass_table_fn: str,
+        reclass_variables: List,
+        variable: Optional[str] = None,
+        fill_method: Optional[str] = None,
+        reproject_method: Optional[str] = None,
+        name: Optional[str] = None,
+        split_dataset: Optional[bool] = True,
+        **kwargs,
+    ) -> List[str]:
+        """
+        This component adds data variable(s) to maps object by reclassifying the data in ``raster_fn`` based on ``reclass_table_fn``.
+
+        Adds model layers:
+
+        * **reclass_variables** maps: reclassified raster data
+
+        Parameters
+        ----------
+        raster_fn: str
+            Source name of raster data in data_catalog. Should be a DataArray. Else use **kwargs to select variables/time_tuple in
+            :py:meth:`hydromt.data_catalog.get_rasterdataset` method.
+        reclass_table_fn: str
+            Source name of reclassification table of `raster_fn` in data_catalog.
+        reclass_variables: list
+            List of reclass_variables from reclass_table_fn table to add to maps. Index column should match values in `raster_fn`.
+        variable: str, optional
+            Name of raster dataset variable to use. This is only required when reading datasets with multiple variables.
+            By default None.
+        fill_method : str, optional
+            If specified, fills nodata values in `raster_fn` using fill_nodata method before reclassifying.
+            Available methods are {'linear', 'nearest', 'cubic', 'rio_idw'}.
+        reproject_method: str, optional
+            See rasterio.warp.reproject for existing methods, by default the data is not reprojected (None).
+        name: str, optional
+            Name of new maps variable, only in case split_dataset=False.
+        split_dataset: bool, optional
+            If data is a xarray.Dataset split it into several xarray.DataArrays (default).
+
+        Returns
+        -------
+        list
+            Names of added model map layers
+        """
+        self.logger.info(
+            f"Preparing map data by reclassifying the data in {raster_fn} based on {reclass_table_fn}"
+        )
+        # Read raster data and remapping table
+        da = self.data_catalog.get_rasterdataset(
+            raster_fn, geom=self.region, buffer=2, **kwargs
+        )
+        if not isinstance(da, xr.DataArray):
+            raise ValueError(
+                f"raster_fn {raster_fn} should be a single variable. "
+                "Please select one using the 'variable' argument"
+            )
+        df_vars = self.data_catalog.get_dataframe(
+            reclass_table_fn, variables=reclass_variables
+        )
+        # Fill nodata
+        if fill_method is not None:
+            ds = ds.raster.interpolate_na(method=fill_method)
+        # Mapping function
+        ds_vars = da.raster.reclassify(reclass_table=df_vars, method="exact")
+        # Reprojection
+        if ds_vars.rio.crs != self.crs and reproject_method is not None:
+            ds_vars = ds_vars.raster.reproject(dst_crs=self.crs)
+        # Add to maps
+        self.set_maps(ds_vars, name=name, split_dataset=split_dataset)
+
+        return list(ds_vars.data_vars.keys())
+
+    # model map
+    @property
+    def maps(self) -> Dict[str, Union[xr.Dataset, xr.DataArray]]:
+        """Model maps. Returns dict of xarray.DataArray or xarray.Dataset"""
+        if len(self._maps) == 0 and self._read:
+            self.read_maps()
+        return self._maps
+
+    def set_maps(
+        self,
+        data: Union[xr.DataArray, xr.Dataset],
+        name: Optional[str] = None,
+        split_dataset: Optional[bool] = True,
+    ) -> None:
+        """Add raster data to the maps component.
+
+        Dataset can either be added as is (default) or split into several
+        DataArrays using the split_dataset argument.
+
+        Arguments
+        ---------
+        data: xarray.Dataset or xarray.DataArray
+            New forcing data to add
+        name: str, optional
+            Variable name, only in case data is of type DataArray or if a Dataset is added as is (split_dataset=False).
+        split_dataset: bool, optional
+            If data is a xarray.Dataset split it into several xarray.DataArrays (default).
+        """
+        data_dict = _check_data(data, name, split_dataset)
+        for name in data_dict:
+            if name in self._maps:
+                self.logger.warning(f"Replacing result: {name}")
+            self._maps[name] = data_dict[name]
+
+    def read_maps(self, fn: str = "maps/*.nc", **kwargs) -> None:
+        """Read model map at <root>/<fn> and add to maps component
+
+        key-word arguments are passed to :py:func:`xarray.open_dataset`
+
+        Parameters
+        ----------
+        fn : str, optional
+            filename relative to model root, may wildcards, by default "maps/*.nc"
+        """
+        self._assert_read_mode
+        ncs = self._read_nc(fn, **kwargs)
+        for name, ds in ncs.items():
+            self.set_maps(ds, name=name)
+
+    def write_maps(self, fn="maps/{name}.nc", **kwargs) -> None:
+        """Write maps to netcdf file at <root>/<fn>
+
+        key-word arguments are passed to :py:meth:`xarray.Dataset.to_netcdf`
+
+        Parameters
+        ----------
+        fn : str, optional
+            filename relative to model root and should contain a {name} placeholder,
+            by default 'maps/{name}.nc'
+        """
+        if len(self._maps) == 0:
+            self.logger.debug("No maps data found, skip writing.")
+        else:
+            self._assert_write_mode
+            self._write_nc(self._maps, fn, **kwargs)
 
     # model geometry files
     @property
     def geoms(self) -> Dict[str, Union[gpd.GeoDataFrame, gpd.GeoSeries]]:
         """Model geometries. Returns dict of geopandas.GeoDataFrame or geopandas.GeoDataSeries
         ..NOTE: previously call staticgeoms."""
-        if not self._geoms:
-            if self._read:
-                self.read_geoms()
+        if not self._geoms and self._read:
+            self.read_geoms()
         return self._geoms
 
     def set_geoms(self, geom: Union[gpd.GeoDataFrame, gpd.GeoSeries], name: str):
@@ -865,6 +1023,7 @@ class Model(object, metaclass=ABCMeta):
         fn : str, optional
             filename relative to model root, may wildcards, by default "geoms/*.nc"
         """
+        self._assert_read_mode
         fns = glob.glob(join(self.root, fn))
         for fn in fns:
             name = basename(fn).split(".")[0]
@@ -882,16 +1041,15 @@ class Model(object, metaclass=ABCMeta):
             filename relative to model root and should contain a {name} placeholder,
             by default 'geoms/{name}.geojson'
         """
-        if not self._write:
-            raise IOError("Model opened in read-only mode")
-        elif not self._geoms:
-            self.logger.warning("No model geoms to write - Exiting")
+        if len(self._geoms) == 0:
+            self.logger.debug("No geoms data found, skip writing.")
             return
+        self._assert_write_mode
         if "driver" not in kwargs:
             kwargs.update(driver="GeoJSON")  # default
         for name, gdf in self._geoms.items():
             if not isinstance(gdf, (gpd.GeoDataFrame, gpd.GeoSeries)) or len(gdf) == 0:
-                self.logger.error(
+                self.logger.warning(
                     f"{name} object of type {type(gdf).__name__} not recognized"
                 )
                 continue
@@ -910,6 +1068,8 @@ class Model(object, metaclass=ABCMeta):
             "The staticgeoms method will be deprecated in future versions, use geoms instead.",
             DeprecationWarning,
         )
+        if not self._geoms and self._read:
+            self.read_staticgeoms()
         self._staticgeoms = self._geoms
         return self._staticgeoms
 
@@ -922,6 +1082,7 @@ class Model(object, metaclass=ABCMeta):
         return self.set_geoms(geom, name)
 
     def read_staticgeoms(self):
+        """This method will be deprecated in future versions, use :py:meth:`~hydromt.Model.read_geoms`"""
         warnings.warn(
             'The read_staticgeoms" method will be deprecated in future versions, use read_geoms instead.',
             DeprecationWarning,
@@ -929,6 +1090,7 @@ class Model(object, metaclass=ABCMeta):
         return self.read_geoms(fn="staticgeoms/*.geojson")
 
     def write_staticgeoms(self):
+        """This method will be deprecated in future versions, use :py:meth:`~hydromt.Model.write_geoms`"""
         warnings.warn(
             'The "write_staticgeoms" method will be deprecated in future versions, use  "write_geoms" instead.',
             DeprecationWarning,
@@ -939,9 +1101,8 @@ class Model(object, metaclass=ABCMeta):
     @property
     def forcing(self) -> Dict[str, Union[xr.Dataset, xr.DataArray]]:
         """Model forcing. Returns dict of xarray.DataArray or xarray.Dataset"""
-        if not self._forcing:
-            if self._read:
-                self.read_forcing()
+        if not self._forcing and self._read:
+            self.read_forcing()
         return self._forcing
 
     def set_forcing(
@@ -977,6 +1138,7 @@ class Model(object, metaclass=ABCMeta):
         fn : str, optional
             filename relative to model root, may wildcards, by default "forcing/*.nc"
         """
+        self._assert_read_mode
         ncs = self._read_nc(fn, **kwargs)
         for name, ds in ncs.items():
             self.set_forcing(ds, name=name)
@@ -992,15 +1154,18 @@ class Model(object, metaclass=ABCMeta):
             filename relative to model root and should contain a {name} placeholder,
             by default 'forcing/{name}.nc'
         """
-        self._write_nc(self._forcing, fn, **kwargs)
+        if len(self._forcing) == 0:
+            self.logger.debug("No forcing data found, skip writing.")
+        else:
+            self._assert_write_mode
+            self._write_nc(self._forcing, fn, **kwargs)
 
     # model state files
     @property
     def states(self) -> Dict[str, Union[xr.Dataset, xr.DataArray]]:
         """Model states. Returns dict of xarray.DataArray or xarray.Dataset"""
-        if not self._states:
-            if self._read:
-                self.read_states()
+        if not self._states and self._read:
+            self.read_states()
         return self._states
 
     def set_states(
@@ -1036,6 +1201,7 @@ class Model(object, metaclass=ABCMeta):
         fn : str, optional
             filename relative to model root, may wildcards, by default "states/*.nc"
         """
+        self._assert_read_mode
         ncs = self._read_nc(fn, **kwargs)
         for name, ds in ncs.items():
             self.set_states(ds, name=name)
@@ -1051,15 +1217,18 @@ class Model(object, metaclass=ABCMeta):
             filename relative to model root and should contain a {name} placeholder,
             by default 'states/{name}.nc'
         """
-        self._write_nc(self._states, fn, **kwargs)
+        if len(self._states) == 0:
+            self.logger.debug("No states data found, skip writing.")
+        else:
+            self._assert_write_mode
+            self._write_nc(self._states, fn, **kwargs)
 
     # model results files; NOTE we don't have a write_results method (that's up to the model kernel)
     @property
     def results(self) -> Dict[str, Union[xr.Dataset, xr.DataArray]]:
         """Model results.  Returns dict of xarray.DataArray or xarray.Dataset"""
-        if not self._results:
-            if self._read:
-                self.read_results()
+        if not self._results and self._read:
+            self.read_results()
         return self._results
 
     def set_results(
@@ -1098,6 +1267,7 @@ class Model(object, metaclass=ABCMeta):
         fn : str, optional
             filename relative to model root, may wildcards, by default "results/*.nc"
         """
+        self._assert_read_mode
         ncs = self._read_nc(fn, **kwargs)
         for name, ds in ncs.items():
             self.set_results(ds, name=name)
@@ -1105,10 +1275,6 @@ class Model(object, metaclass=ABCMeta):
     def _write_nc(
         self, nc_dict: Dict[str, Union[xr.DataArray, xr.Dataset]], fn, **kwargs
     ) -> None:
-        if not self._write:
-            raise IOError("Model opened in read-only mode")
-        if len(nc_dict) == 0:
-            self.logger.info("No data found - exiting")
         for name, ds in nc_dict.items():
             if not isinstance(ds, (xr.Dataset, xr.DataArray)) or len(ds) == 0:
                 self.logger.error(
@@ -1359,22 +1525,3 @@ def _check_equal(a, b, name="") -> Dict[str, str]:
     except AssertionError as e:
         errors.update({name: e})
     return errors
-
-
-class AuxmapsModel(AuxmapsMixin, Model):
-    def __init__(
-        self,
-        root: str = None,
-        mode: str = "w",
-        config_fn: str = None,
-        data_libs: List[str] = None,
-        logger=logger,
-    ):
-        # Initialize with the Model class
-        super().__init__(
-            root=root,
-            mode=mode,
-            config_fn=config_fn,
-            data_libs=data_libs,
-            logger=logger,
-        )