+++ tests/conftest.py
@@ -1,14 +1,16 @@
 import pytest
+from os.path import join, dirname, abspath
 import numpy as np
 import pandas as pd
 import geopandas as gpd
 import xarray as xr
-from shapely.geometry import box
 
-from hydromt import Model, GridModel, LumpedModel, NetworkModel
-from hydromt.models.model_api import AuxmapsModel
+from hydromt import Model, GridModel, LumpedModel, NetworkModel, MODELS
+from hydromt.data_catalog import DataCatalog
 from hydromt import raster, vector, gis_utils
 import pyflwdir
+
+# DATADIR = join(dirname(abspath(__file__)), "data")
 
 
 @pytest.fixture
@@ -51,7 +53,7 @@ def df_time():
 @pytest.fixture
 def geodf(df):
     gdf = gpd.GeoDataFrame(
-        data=df.drop(columns=["longitude", "latitude"]),
+        data=df.copy().drop(columns=["longitude", "latitude"]),
         geometry=gpd.points_from_xy(df["longitude"], df["latitude"]),
         crs=4326,
     )
@@ -68,7 +70,7 @@ def world():
 def ts(geodf):
     dates = pd.date_range("01-01-2000", "12-31-2000", name="time")
     ts = pd.DataFrame(
-        index=geodf.index,
+        index=geodf.index.values,
         columns=dates,
         data=np.random.rand(geodf.index.size, dates.size),
     )
@@ -156,24 +158,33 @@ def demuda():
 
 
 @pytest.fixture
+def griduda():
+    import xugrid as xu
+
+    bbox = [12.09, 46.49, 12.10, 46.50]  # Piava river
+    data_catalog = DataCatalog(data_libs=["artifact_data"])
+    da = data_catalog.get_rasterdataset("merit_hydro", bbox=bbox, variables="elevtn")
+    gdf_da = da.raster.vector_grid()
+    gdf_da["value"] = da.values.flatten()
+    gdf_da.index.name = "mesh2d_nFaces"
+    uda = xu.UgridDataset.from_geodataframe(gdf_da)
+    uda = uda["value"]
+    uda.name = "elevtn"
+
+    return uda
+
+
+@pytest.fixture
 def model(demda, world, obsda):
     mod = Model()
     mod.setup_region({"geom": demda.raster.box})
     mod.setup_config(**{"header": {"setting": "value"}})
     mod.set_staticmaps(demda, "elevtn")  # will be deprecated
     mod.set_geoms(world, "world")
+    mod.set_maps(demda, "elevtn")
     mod.set_forcing(obsda, "waterlevel")
     mod.set_states(demda, "zsini")
     mod.set_results(obsda, "zs")
-    return mod
-
-
-@pytest.fixture
-def auxmap_model(demda):
-    mod = AuxmapsModel()
-    mod.setup_region({"geom": demda.raster.box})
-    mod.setup_config(**{"header": {"setting": "value"}})
-    mod.set_auxmaps(demda, "elevtn")
     return mod
 
 
@@ -213,11 +224,9 @@ def network_model():
 
 @pytest.fixture
 def mesh_model(demuda):
-    from hydromt import MeshModel
-
-    mod = MeshModel()
+    mod = MODELS.load("mesh_model")()
     # region = gpd.GeoDataFrame(
-    #     geometry=[box(*demuda.ugrid.grid.bounds)], crs=demuda.ugrid.crs
+    #     geometry=[box(*demuda.ugrid.grid.bounds)], crs=demuda.ugrid.grid.crs
     # )
     # mod.setup_region({"geom": region})
     mod.setup_config(**{"header": {"setting": "value"}})